src/eval_fid.py
"""
FID evaluation script (simplified version).
For full FID, you'd need to install pytorch-fid or use a library.
This is a placeholder that computes basic statistics.
"""
import os
import yaml
import argparse
import torch
import numpy as np
from torchvision import transforms
from PIL import Image

def compute_statistics(images):
    """Compute mean and covariance statistics for FID (simplified)."""
    # Flatten images
    images = images.reshape(images.shape[0], -1)
    mu = np.mean(images, axis=0)
    sigma = np.cov(images, rowvar=False)
    return mu, sigma

def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--cfg", type=str, default="configs/base.yaml")
    parser.add_argument("--ckpt", type=str, required=True)
    parser.add_argument("--n_samples", type=int, default=1000)
    args = parser.parse_args()
    
    cfg = yaml.safe_load(open(args.cfg, "r"))
    device = "cuda" if torch.cuda.is_available() else "cpu"
    
    # Load model
    ckpt = torch.load(args.ckpt, map_location=device)
    from unet import TinyUNet
    from ddpm import DDPM
    
    num_classes = len(ckpt["classes"])
    model = TinyUNet(num_classes=num_classes, base_ch=cfg["base_ch"]).to(device)
    model.load_state_dict(ckpt.get("ema_model", ckpt["model"]))
    model.eval()
    
    ddpm = DDPM(timesteps=cfg["timesteps"], device=device)
    
    # Generate samples
    print(f"Generating {args.n_samples} samples...")
    all_samples = []
    batch_size = 64
    
    with torch.no_grad():
        for i in range(0, args.n_samples, batch_size):
            n = min(batch_size, args.n_samples - i)
            # Sample random classes
            y = torch.randint(0, num_classes, (n,), device=device)
            samples = ddpm.sample(
                model, n=n, img_size=cfg["img_size"], 
                y=y, seed=42 + i
            )
            all_samples.append(samples.cpu())
    
    all_samples = torch.cat(all_samples, dim=0)
    print(f"Generated {all_samples.shape[0]} samples")
    print(f"Sample shape: {all_samples.shape}")
    print(f"Sample range: [{all_samples.min():.3f}, {all_samples.max():.3f}]")
    
    # Save statistics
    stats = compute_statistics(all_samples.numpy())
    print(f"Mean shape: {stats[0].shape}")
    print(f"Covariance shape: {stats[1].shape}")
    
    # For full FID, you'd compare these stats with real data stats
    print("\nNote: For full FID score, install pytorch-fid or use a proper FID library.")
    print("This script only computes basic statistics.")

if __name__ == "__main__":
    main()
